Austyn: Fixed uint casting

+++ Dump.cs
@@ -4,8 +4,8 @@ public class Dump : IInstruction{
 	public Dump(){}
 	public uint Val{
 		get{
-			uint retVal = 14 << 28; //1110
-			return retVal;
+			int retVal = 14 << 28; //1110
+			return (uint)retVal;
 		}
 	}
 }
